requirements.txt
constants==0.6.0
ibapi==9.81.1.post1
numpy==1.20.3
pandas==1.3.0
psycopg2_binary==2.9.2
PyYAML==6.0
setuptools==59.6.0